Update the intersects logic to cover both the before and after
This commit makes two important decisions:

When finding hunk intersections, it requires both the before and after to match for it to be considered intersected. I made this choice because it is concevable to have hunks shift in a way that we accidentally pick up two overlaps and cause strange unassigning behaviours. By requiring both the before and after ranges to overlap, it does mean that we might loose some assignments, but this initially seems like the better comprimise.

Behaviour around how you consider a range of 0 lines “intersecting”. I ended up deciding that if the “start” number is inside the other range, even if the line count is 0, it is considered intersecting.