Improve worktreeChanges function
This commit contains two changes.

It adapts the function to better represent the intended behaviour, now if you pass a stackId it will still return any unassigned selections

Sorts hunk headers. Hunk headers need to be ordered from the top of a file to the bottom. This adds in an ordering function to help us do that