A first shot at producing stacks from a graph
Also: remove CommitDetails for better performance and less memory consumption.

Make it minimal, leave it to shortly before the UI for details.
Otherwise, it's easy to run out of memory with big graphs unless care is taken.
Ideally we are just lazy enough to not have to take care.

Remove commit details for more a little speed

Any waste is bad when dealing with a lot of commits, and we don't need
it for anything yet.

Also: set graph to StableGraph implementation for stable node indices.
This is beneficial as it allows us to see the order of segment creation,
while having them for visualization early on as well.