Switch everyone to v3
# Default v3 enabled
- We should default all users to v3.
- We should inform users that have just been moved that they can switch back for now.
- Need to do a hard reload after the switch.
We probably want the frontend to take responsibility for the switch. That way we can switch at a good time (after onboarding) and correctly reload the frontend.

We can:

- Use a persisted value "switched_to_v3" that defaults to false
- When the frontend is loaded on the v2 version and "switched_to_v3" is false
- We switch to v3 & hard reload
- Show the user the modal with the option to switch back to v2.
- Set "switched_to_v3" to true
- If "switched_to_v3" is true, we won't prompt the user again.
I have an `onMount` in the component that does the switch over to v3.

- We perform a hard reload in the switching logic which means we need to be a little clever about how to actually get the modal to show.
- By performing a hard reload we have no way to refer to the *new* modal component that is rendered after the reload.
- As such, we need to persist the fact that we should show the modal.


I also added a keybind where you can type “debugswitch” which sets the client back to v2 and allows the switch to happen again.