Propagate the limit to extend walk more naturally
We can be in a situation where there is no goal to persue, yet we
hit a commit that was already seen.

This previously made the traversal stop even if the commit was discovered
by a tip that was limited more, leading to traversals that were too short.

Now the possibly larger limit is assigned to the tips that found the commit
in question, extending the search.