fix: use correct symbol for referencing branches on forges
Use the correct symbol for referencing branches on different forges. The
issue arose when using stacked branches on GitLab in a merge request. We
used `#` instead of `!` which worked fine for GitHub but did not work 
for GitLab. 

Now, we utilize the already present information in the `ForgePrService` 
such that we always use the correct symbol when referencing branches.