Let `changes_in_branch()` use the new data structures
That way it is more robust and doesn't depend on stack-ids anymore,
just on a workspace view.

Note that in order to do that, I also ~~decided~~ had toto do perform a long-standing
refactoring that makes the legacy-types independent of the old code-base,
while linking both types through the type-system to zero the chance for drift.
That way it's possible to use `but-graph` in `command-context`, so that
there can be utility functions that create metadata and the graph as well.