feat: remove resizer border imitation and drop resizer highlight (#9806)
* Made details view pop

Rearrange div structure to achieve an unfolded lane layout. 
The idea is to adjust the right margin and render the details 
view as an absolutely positioned element, allowing the 
scrollbar to be on the right side of the whole lane.

Otherwise, the scrollbar would appear in the middle between 
sections, which doesn’t fit the layout.

* Details view resizer refactoring

Because we decided not to resize the commit/branch details 
section, and the changes preview section now takes only the 
maximum height of the remaining space or fits the content, 
basically only one section is resizable — the changes list. 
That’s why we don’t need complex calculations here.

* feat: remove resizer border imitation and drop resizer highlight