Avoid using the Git bash environment on Windows
The problem here is that the paths in these envirnoment variables
will work in Git bash, and not on the normal OS. This can lead
to difficult-to-debug errors as now it matters *where* the path
is used.

Those who use a Git bash by default on Windows will not have a good time.