feat: show forge integration banner with config prompt (#10294)
* feat: show forge integration banner with config prompt

This introduces logic to detect if the user can configure a GitHub or GitLab integration. If this is the case, the UI displays a banner in the relevant view, prompting users to configure the integration and providing documentation links. Adds supporting logic for dismissing the prompt and utility helpers for forge names/labels/links.

Also updates the usage in certain components to leverage the new detection and display workflow, and improves workspace explanation in project setup.

* InfoMessage: remove "pop" style

Update InfoMessage.svelte

* update styles for the Forge banner

* update "target" onboarding screen

* fix lint errors

* Update UnassignedViewForgePrompt.svelte

---------

Co-authored-by: Pavel Laptev <pawellaptew@gmail.com>